Classification_App/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.0
numpy==1.22.2
Werkzeug==2.1.1
zipp==3.8.0
torch==1.10.2+cpu
torchaudio==0.10.2+cpu
torchvision==0.11.3+cpu
gunicorn==20.1.0